+++ src/drivers/EnvironmentDriver.ts
@@ -1,5 +1,5 @@
 import { SyncFeatureManagerDriver } from '../base-drivers/SyncFeatureManagerDriver'
-import { CommonValueParams } from '../types/common.types'
+import { CommonValueParams, ValueReturnType } from '../types/common.types'
 
 /**
  * Uses process.env to get the values. Supports sync and async operations.
@@ -8,18 +8,22 @@ export class EnvironmentDriver<
   Flags extends Record<string, any> = Record<string, any>,
   Context = never,
 > extends SyncFeatureManagerDriver<Flags, Context> {
-  async getRawValue<K extends string & keyof Flags>(
-    key: K,
-    params?: CommonValueParams<Flags, K>
-  ): Promise<Flags[K] | null> {
+  async getRawValue<
+    K extends string & keyof Flags,
+    Params extends CommonValueParams<Flags, K> | undefined = undefined,
+  >(key: K, params?: Params): Promise<ValueReturnType<Flags, K, Params>> {
     return this.getRawValueSync(key, params)
   }
 
-  getRawValueSync<K extends string & keyof Flags>(
-    key: K,
-    params?: CommonValueParams<Flags, K>
-  ): Flags[K] | null {
-    return (process.env[key] as Flags[K]) ?? params?.defaultValue ?? null
+  getRawValueSync<
+    K extends string & keyof Flags,
+    Params extends CommonValueParams<Flags, K> | undefined = undefined,
+  >(key: K, params?: Params): ValueReturnType<Flags, K, Params> {
+    return (
+      (process.env[key] as Flags[K]) ??
+      params?.defaultValue ??
+      (null as ValueReturnType<Flags, K, Params>)
+    )
   }
 
   getAllRawValuesSync(): Flags {