+++ src/managers/AsyncFeatureManager.ts
@@ -1,5 +1,5 @@
 import { AsyncFeatureManagerDriver } from '../base-drivers/AsyncFeatureManagerDriver'
-import { CommonValueParams } from '../types/common.types'
+import { CommonValueParams, ValueReturnType } from '../types/common.types'
 import { IAsyncFeatureManager } from '../types/IAsyncFeatureManager'
 
 /**
@@ -18,10 +18,10 @@ export class AsyncFeatureManager<Flags extends Record<string, any>, Context>
     this.driver = driver
   }
 
-  async getValue<K extends string & keyof Flags>(
-    key: K,
-    params?: CommonValueParams<Flags, K>
-  ): Promise<Flags[K] | null> {
+  async getValue<
+    K extends string & keyof Flags,
+    Params extends CommonValueParams<Flags, K> | undefined = undefined,
+  >(key: K, params?: Params): Promise<ValueReturnType<Flags, K, Params>> {
     return this.driver.getValue(key, params)
   }
 
@@ -29,10 +29,10 @@ export class AsyncFeatureManager<Flags extends Record<string, any>, Context>
     return this.driver.getAllValues(params)
   }
 
-  async getRawValue<K extends string & keyof Flags>(
-    key: K,
-    params?: CommonValueParams<Flags, K>
-  ): Promise<Flags[K] | null> {
+  async getRawValue<
+    K extends string & keyof Flags,
+    Params extends CommonValueParams<Flags, K> | undefined = undefined,
+  >(key: K, params?: Params): Promise<ValueReturnType<Flags, K, Params>> {
     return this.driver.getRawValue(key, params)
   }
 
@@ -40,17 +40,17 @@ export class AsyncFeatureManager<Flags extends Record<string, any>, Context>
     return this.driver.getAllRawValues(params)
   }
 
-  async assertGetValue<K extends string & keyof Flags>(
-    key: K,
-    params?: CommonValueParams<Flags, K>
-  ): Promise<Flags[K]> {
+  async assertGetValue<
+    K extends string & keyof Flags,
+    Params extends CommonValueParams<Flags, K> | undefined = undefined,
+  >(key: K, params?: Params): Promise<ValueReturnType<Flags, K, Params>> {
     return this.driver.assertGetValue(key, params)
   }
 
-  async assertGetRawValue<K extends string & keyof Flags>(
-    key: K,
-    params?: CommonValueParams<Flags, K>
-  ): Promise<Flags[K]> {
+  async assertGetRawValue<
+    K extends string & keyof Flags,
+    Params extends CommonValueParams<Flags, K> | undefined = undefined,
+  >(key: K, params?: Params): Promise<ValueReturnType<Flags, K, Params>> {
     return this.driver.assertGetRawValue(key, params)
   }