+++ src/types/common.types.ts
@@ -1,9 +1,6 @@
 /**
  * Common optional parameters for retrieving a flag.
  */
-/**
- * Common optional parameters for retrieving a flag.
- */
 export type CommonValueParams<Flags, K extends keyof Flags> = {
   /**
    * The default value to use if the flag is not found.
@@ -14,3 +11,13 @@ export type CommonValueParams<Flags, K extends keyof Flags> = {
    */
   context?: any
 }
+
+export type ValueReturnType<
+  Flags,
+  K extends keyof Flags,
+  Params extends CommonValueParams<Flags, K> | undefined = undefined,
+> = Params extends {
+  defaultValue: infer DefaultValue
+}
+  ? Flags[K] | DefaultValue | null | undefined
+  : Flags[K] | null | undefined